+++ src/index.ts
@@ -1,4 +1,4 @@
-export { useForm } from "./useForms";
+export { useForm } from "./useForm";
 export { useFields } from "./useFields";
 export { useErrors } from "./useErrors";
 export {